feat(T16): tokens columns, delegated to itok
The `tokens` column in `scan` and `log` has been an explicit null since the
scaffold, holding the shape open for this. It is now filled by `itok
estimate --bpe`, so the counts are o200k and TOKENIZER-backed -- measured
and deterministic, never a model's guess and never bytes/4 (section C).
Not reimplemented here: two estimators disagreeing is the same defect as
two rule sets, and the arithmetic kind argues loudest for itself because
both numbers look plausible (V8).

BATCHED, AND THE MEASUREMENT DECIDED IT. `itok estimate` counts FILES and
a statement is a span inside one, so the statements are written to a
private directory and counted in ONE call. Measured on an Apple M4, 10
core / 16GB, 200 statements: one process per statement via `/dev/stdin`
costs 57ms each -- 11 seconds per scan -- against 74ms for all 200 in a
single call. The cost is the BPE table load, paid once per spawn.

The numbers carry the box they were taken on, because a timing without its
conditions is an assertion wearing a measurement's clothes -- and that box
is the FAST one. The development box is roughly 5x slower, which puts the
per-statement shape near 57 SECONDS for one scan, past the point section C
names where a step becomes one people learn to bypass. The ratio is what
decides this, not the absolute figures.

PRIVACY OF THE SCRATCH. Mode 0700 on the directory and 0600 on each file,
set rather than inherited -- `create` honours the umask, commonly 022. The
directory removes itself through `Drop` rather than a cleanup call, because
the counting path has three fallible steps after it exists and an early `?`
walks straight past a cleanup someone has to remember. V15 forbids EGRESS,
a byte reaching a host, and this reaches none: the bytes were already on
this disk in the corpus file they were read from.

A CONCURRENCY BUG THE SUITE FOUND. The directory was named by PID alone,
and `cargo test` runs many scans at once in one process, so parallel counts
deleted each other's files mid-count. The symptom was a column that went
missing under load and never when run alone. It now carries a per-call
counter, and two tests hold the property -- one on the names, one running
four counts in parallel.

V26 THROUGHOUT. `itok` arrives on PATH, not through Cargo, so its absence
is a SKIP that is NAMED: the column shows `-` and a line says how to get it
back. The verb still reports. The tests, by contrast, assert
UNCONDITIONALLY -- the dev shell provides `itok` and the gate already
depends on it, so a "skip if absent" guard there would be a branch that
never runs, and on a box where it did run it would turn a real failure into
a silent pass.

Aligned by FILENAME rather than output order: relying on a sibling to
answer in the order it was asked is depending on a promise it never made,
and the failure would be silent, pairing each statement with its
neighbour's count.

`scan::run` now filters rows and their texts TOGETHER through one
predicate. Two passes would be one predicate in two places, and `--top 5`
would carry the first five texts of a different five rows.

380 tests, gate green, coverage 99.16 against the 99.26 floor.

+++ src/log.rs
@@ -28,12 +28,14 @@ pub struct Entry {
     pub label: String,
     pub artifact: String,
     pub fires: u64,
-    /// Tokens reclaimed. `None` until `itok` is wired in (V8, T16).
+    /// Tokens RECLAIMED -- what this statement cost the window before it
+    /// was moved out. Delegated to `itok` (V8) and filled by the caller,
+    /// so this module stays a pure function of the ledger.
     ///
-    /// NULL rather than a guess. This crate does not own token accounting,
-    /// and a char/4 stand-in printed in the same column a tokenizer will
-    /// later fill is a number that looks measured and is not.
-    pub tokens: Option<u64>,
+    /// NULL rather than a guess when the sibling is absent (V26). A
+    /// char/4 stand-in printed in the column a tokenizer fills is a
+    /// number that looks measured and is not.
+    pub tokens: Option<u32>,
     /// Unix seconds, as recorded.
     pub at: u64,
 }
